fix: Int types was changed to uint8_t

+++ gpioHandler.c
@@ -2,10 +2,10 @@
 
 #define MS 1000
 
-static int ports[11] = {0,0,0,0,0,0,0,0,0,0,0};
+static uint8_t ports[11] = {0,0,0,0,0,0,0,0,0,0,0};
 
-void delay(int ms){
-	for(volatile int i = 0; i < ms*MS; i++);
+void delay(uint8_t ms){
+	for(volatile uint8_t i = 0; i < ms*MS; i++);
 }
 
 void writeHigh(Pin* pin){
@@ -20,14 +20,14 @@ void togglePin(Pin* pin){
 	pin->port->ODR ^= 1 << pin->index;
 }
 
-void virtualPwmWrite(Pin* pin, int value){
+void virtualPwmWrite(Pin* pin, uint8_t value){
 	writeHigh(pin);
-	for(int i = 0; i < value*MS/10; i++);
+	for(uint8_t i = 0; i < value*MS/10; i++);
 	writeLow(pin);
-	for(int i = 0; i < (100 - value)*MS/10; i++);
+	for(uint8_t i = 0; i < (100 - value)*MS/10; i++);
 }
 
-void pinStart(Pin *pin, GPIO_TypeDef* port, int index, int mode){
+void pinStart(Pin *pin, GPIO_TypeDef* port, uint8_t index, uint8_t mode){
 	index = ((uint32_t)port - GPIOA_BASE) / (GPIOB_BASE - GPIOA_BASE);
 
 	if(ports[index] != 1) {
@@ -52,7 +52,7 @@ void setPullUp(Pin* pin){
 	pin->port->PUPDR |= 1 << pin->index*2;
 }
 
-int readPin(Pin* pin) {
+uint8_t readPin(Pin* pin) {
 	if(((pin->port->IDR) & (1 << pin->index)) != 0)
 		return 1;
 	else return 0;
